test: isolate functional coverage job override
