+++ cmd/gocoveragecheck/main.go
@@ -204,7 +204,7 @@ func parseConfig() config {
 	flag.StringVar(&cfg.coverpkg, "coverpkg", "", "comma-separated import paths to measure; defaults to backend-owned packages")
 	flag.StringVar(&cfg.functionalQuarantine, "functional-quarantine", "", "strict functional quarantine JSON manifest; discovers and subtracts its package/test selectors")
 	flag.IntVar(&cfg.jobs, "jobs", 0, "maximum concurrent go test packages; defaults to runtime CPU count for non-Windows unit coverage, 1 for Windows unit coverage, and 2 for functional coverage")
-	flag.IntVar(&cfg.testJobsOverride, "test-jobs", 0, "positive instrumented go test -p override; omitted uses -jobs (functional subprocess I/O wait can benefit from controlled oversubscription)")
+	registerTestJobsFlag(&cfg)
 	flag.StringVar(&cfg.generateManifest, "generate-manifest", "", "create a deterministic package-minimum manifest from this lane's coverage profile")
 	flag.StringVar(&cfg.updateManifest, "update-manifest", "", "update an existing package-minimum manifest from a complete compatible profile sample set")
 	flag.StringVar(&cfg.updateProfiles, "update-profiles", "", "comma-separated complete coverage profiles for -update-manifest; requires at least five compatible profiles")
@@ -228,17 +228,13 @@ func parseConfig() config {
 	flag.DurationVar(&cfg.timeout, "timeout", 5*time.Minute, "go test timeout")
 	flag.BoolVar(&cfg.totalOnly, "total-only", false, "disable package-local coverage gates while retaining per-package reporting")
 	flag.Parse()
-	flag.Visit(func(visited *flag.Flag) {
-		if visited.Name == "test-jobs" {
-			cfg.testJobsOverrideSet = true
-		}
-	})
+	markTestJobsOverrideSet(&cfg)
 	return cfg
 }
 
 func validateConfig(cfg config) error {
-	if cfg.testJobsOverride < 0 || (cfg.testJobsOverride == 0 && cfg.testJobsOverrideSet) {
-		return fmt.Errorf("configure go coverage: -test-jobs must be a positive integer (got %d)", cfg.testJobsOverride)
+	if err := validateTestJobsOverride(cfg); err != nil {
+		return err
 	}
 	manifestOperations := 0
 	for _, value := range []string{cfg.generateManifest, cfg.updateManifest, cfg.packageManifest} {
@@ -393,16 +389,6 @@ func (cfg config) testJobs(targetOS string, logicalCPUs int) int {
 		// usable count. runtime.NumCPU normally guarantees a positive value.
 	}
 	return defaultCoverageJobs
-}
-
-// instrumentedTestJobs keeps the instrumented test window independently
-// tunable: functional subprocesses spend time waiting on I/O, so controlled
-// oversubscription can reduce wall time without changing discovery's jobs.
-func (cfg config) instrumentedTestJobs(targetOS string, logicalCPUs int) int {
-	if cfg.testJobsOverride > 0 {
-		return cfg.testJobsOverride
-	}
-	return cfg.testJobs(targetOS, logicalCPUs)
 }
 
 func packageCoverageBaselinePackages(cfg config, repoRoot string) (map[string]struct{}, error) {
